requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
apscheduler==3.11.0
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
frozenlist==1.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
magic-filter==1.0.12
multidict==6.6.4
numpy==2.3.3
outcome==1.3.0.post0
pandas==2.3.2
propcache==0.3.2
pydantic==2.11.9
pydantic-core==2.33.2
pysocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
selenium==4.21.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sqlalchemy==2.0.43
trio==0.31.0
trio-websocket==0.12.2
typing-extensions==4.15.0
typing-inspection==0.4.1
tzdata==2025.2
tzlocal==5.3.1
undetected-chromedriver==3.5.5
urllib3==2.5.0
websockets==15.0.1
wsproto==1.2.0
yarl==1.20.1
